Added OpenAPI config and updated ClientService functionality
The commit adds a new OpenAPI configuration file, OpenAPIConfig.java, to support API documentation. It also updates the ClientServiceI.java to add a new method `findBySharedKey(String email)`, allowing clients to be found by a shared key. The pom.xml file is updated to include the necessary dependencies for OpenAPI and OpenCSV. The ClientService.java is updated to provide the implementation of the `findBySharedKey()` method. Additionally, a new API endpoint `v1/client/sharedKey/{sharedKey}` is added to the ClienteController.java to enable retrieving clients by a shared key. Lastly, the application's property file and .gitignore have been updated as well.